llm_bench: Move resolve_kokoro_speaker_embedding into tts_utils
